mvcc/reader: address comments
